changed type of entrance fee

+++ DB Scripts/create table specialEvent.sql
@@ -1,7 +1,7 @@
 create table specialEvent(
 	eventDATE date,
 	name char(30), 
-	entranceFee  integer, 
+	entranceFee  decimal(5,2),
 	startTIME char(20), 
 	endTIME char(20),
 	capacity number(5),
